Rename Exception to InvalidConfigurationException
